Use single spinner component across app
