--- packages/graph-explorer/src/components/LoadingSpinner/index.ts
@@ -1,4 +1,0 @@
-export { default } from "./LoadingSpinner";
-export { default as LoaderIcon } from "@/components/icons/LoaderIcon";
-export type { LoadingSpinnerProps } from "./LoadingSpinner";
-export * from "./LoadingSpinner";
